+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Function `Random::replace_candidate()` is slightly faster
 
+### Removed
+- Removed Catch2 submodule
+
 ## [1.0.0] - 2019-08-13
 ### Added
 - Cache class